Remove some unused code of `ExecutionsTDNode`
